Adding Client Server Classes
Extracted the code for the submission in a class to be able to use it
with the server or the client.
