+++ TP1/Tp1/Tp1/Submission.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Tp1
+{
+    class Submission
+    {
+        public void submit(String filename)
+        {
+            byte[] bytes = System.IO.File.ReadAllBytes(filename);
+        }
+    }
+}
